added BaseSyncActivity class
