+++ services_app/src/androidTest/java/org/opendatakit/activites/SyncActivity/AuthenticatedUserStateTest.java
@@ -15,8 +15,6 @@ import static org.hamcrest.Matchers.instanceOf;
 import static org.hamcrest.Matchers.is;
 import static org.hamcrest.Matchers.not;
 
-import android.app.Activity;
-import android.content.Context;
 import android.content.Intent;
 
 import androidx.test.core.app.ActivityScenario;
@@ -25,9 +23,7 @@ import androidx.test.espresso.intent.Intents;
 import androidx.test.espresso.intent.matcher.IntentMatchers;
 import androidx.test.espresso.matcher.RootMatchers;
 import androidx.test.espresso.matcher.ViewMatchers;
-import androidx.test.platform.app.InstrumentationRegistry;
 
-import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 import org.opendatakit.consts.IntentConsts;
@@ -47,14 +43,10 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.Random;
 
-public class AuthenticatedUserStateTest {
-
-    private final String TEST_SERVER_URL = "https://testUrl.com";
+public class AuthenticatedUserStateTest extends BaseSyncActivity {
 
     private final String TEST_USERNAME = "testUsername";
     private final String TEST_PASSWORD = "testPassword";
-
-    private ActivityScenario<SyncActivity> activityScenario;
 
     @Before
     public void setUp() {
@@ -111,8 +103,8 @@ public class AuthenticatedUserStateTest {
 
     @Test
     public void verifyChangeSyncTypeTest() {
-        String [] syncTypes = getContext().getResources().getStringArray(R.array.sync_attachment_option_names);
-        String type=syncTypes[new Random().nextInt(4)];
+        String[] syncTypes = getContext().getResources().getStringArray(R.array.sync_attachment_option_names);
+        String type = syncTypes[new Random().nextInt(4)];
 
         onView(withId(R.id.autoInputSyncType)).perform(ViewActions.click());
         onData(allOf(is(instanceOf(String.class)), is(type)))
@@ -201,32 +193,4 @@ public class AuthenticatedUserStateTest {
         onView(withId(R.id.tvSignInWarnHeadingSync)).check(matches(isDisplayed()));
         onView(withId(R.id.btnDrawerLogin)).check(matches(withText(getContext().getString(R.string.drawer_sign_in_button_text))));
     }
-
-    @After
-    public void clearTestEnvironment() {
-        activityScenario.onActivity(activity -> {
-            PropertiesSingleton props = activity.getProps();
-            assertThat(props).isNotNull();
-
-            Map<String, String> serverProperties = UpdateServerSettingsFragment.getUpdateUrlProperties(
-                    activity.getString(org.opendatakit.androidlibrary.R.string.default_sync_server_url)
-            );
-            assertThat(serverProperties).isNotNull();
-            serverProperties.put(CommonToolProperties.KEY_FIRST_LAUNCH,"true");
-            serverProperties.put(CommonToolProperties.KEY_SYNC_ATTACHMENT_STATE, null);
-            props.setProperties(serverProperties);
-        });
-        activityScenario.close();
-    }
-
-    private Activity getActivity() {
-        final Activity[] activity1 = new Activity[1];
-        activityScenario.onActivity(activity -> activity1[0] =activity);
-        return activity1[0];
-    }
-
-    private Context getContext() {
-        return InstrumentationRegistry.getInstrumentation().getTargetContext();
-    }
-
 }
