+++ services_app/src/androidTest/java/org/opendatakit/activites/SyncActivity/GeneralStateTest.java
@@ -9,7 +9,6 @@ import static androidx.test.espresso.matcher.ViewMatchers.withId;
 import static androidx.test.espresso.matcher.ViewMatchers.withText;
 import static com.google.common.truth.Truth.assertThat;
 
-import android.content.Context;
 import android.content.Intent;
 
 import androidx.test.core.app.ActivityScenario;
@@ -18,9 +17,7 @@ import androidx.test.espresso.action.ViewActions;
 import androidx.test.espresso.intent.Intents;
 import androidx.test.espresso.intent.matcher.IntentMatchers;
 import androidx.test.espresso.matcher.RootMatchers;
-import androidx.test.platform.app.InstrumentationRegistry;
 
-import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 import org.opendatakit.consts.IntentConsts;
@@ -29,17 +26,12 @@ import org.opendatakit.properties.PropertiesSingleton;
 import org.opendatakit.services.R;
 import org.opendatakit.services.preferences.activities.AppPropertiesActivity;
 import org.opendatakit.services.sync.actions.activities.SyncActivity;
-import org.opendatakit.services.sync.actions.activities.VerifyServerSettingsActivity;
 import org.opendatakit.services.sync.actions.fragments.UpdateServerSettingsFragment;
 
 import java.util.Collections;
 import java.util.Map;
 
-public class GeneralStateTest {
-
-    private final String TEST_SERVER_URL = "https://testUrl.com";
-
-    private ActivityScenario<SyncActivity> activityScenario;
+public class GeneralStateTest extends BaseSyncActivity {
 
     @Before
     public void setUp() {
@@ -121,25 +113,4 @@ public class GeneralStateTest {
         onView(withId(org.opendatakit.androidlibrary.R.id.versionText)).check(matches(isDisplayed()));
         btnAboutUs.check(matches(isNotEnabled()));
     }
-
-    @After
-    public void clearTestEnvironment() {
-        activityScenario.onActivity(activity -> {
-            PropertiesSingleton props = activity.getProps();
-            assertThat(props).isNotNull();
-
-            Map<String, String> serverProperties = UpdateServerSettingsFragment.getUpdateUrlProperties(
-                    activity.getString(org.opendatakit.androidlibrary.R.string.default_sync_server_url)
-            );
-            assertThat(serverProperties).isNotNull();
-            serverProperties.put(CommonToolProperties.KEY_FIRST_LAUNCH,"true");
-            props.setProperties(serverProperties);
-        });
-        activityScenario.close();
-    }
-
-    private Context getContext() {
-        return InstrumentationRegistry.getInstrumentation().getTargetContext();
-    }
-
 }
